src/requirements.txt
aiosmtpd
pymongo
mysql-connector
